+++ Branch-SDK/src/io/branch/referral/Branch.java
@@ -55,7 +55,6 @@ import io.branch.indexing.BranchUniversalObject;
 import io.branch.indexing.ContentDiscoverer;
 import io.branch.referral.network.BranchRemoteInterface;
 import io.branch.referral.util.CommerceEvent;
-import io.branch.referral.util.InstantAppUtil;
 import io.branch.referral.util.LinkProperties;
 
 /**
@@ -3643,7 +3642,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
                 }
             }
         }
-        return doShowInstallPrompt(activity, requestCode, installReferrerString);
+        return InstantAppUtil.doShowInstallPrompt(activity, requestCode, installReferrerString);
     }
 
     /**
@@ -3657,7 +3656,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
      */
     public static boolean showInstallPrompt(@NonNull Activity activity, int requestCode, @Nullable String referrer) {
         String installReferrerString = Defines.Jsonkey.IsFullAppConv.getKey() + "=true&" + referrer;
-        return doShowInstallPrompt(activity, requestCode, installReferrerString);
+        return InstantAppUtil.doShowInstallPrompt(activity, requestCode, installReferrerString);
     }
 
     /**
@@ -3683,29 +3682,6 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
             }
         }
         return false;
-    }
-
-
-    private static boolean doShowInstallPrompt(@NonNull Activity activity, int requestCode, @Nullable String referrer) {
-        if (activity == null) {
-            Log.e("BranchSDK", "Unable to show install prompt. Activity is null");
-            return false;
-        } else if (!isInstantApp(activity)) {
-            Log.e("BranchSDK", "Unable to show install prompt. Application is not an instant app");
-            return false;
-        } else {
-            Intent intent = (new Intent("android.intent.action.VIEW")).setPackage("com.android.vending").addCategory("android.intent.category.DEFAULT")
-                    .putExtra("callerId", activity.getPackageName())
-                    .putExtra("overlay", true);
-            Uri.Builder uriBuilder = (new Uri.Builder()).scheme("market").authority("details").appendQueryParameter("id", activity.getPackageName());
-            if (!TextUtils.isEmpty(referrer)) {
-                uriBuilder.appendQueryParameter("referrer", referrer);
-            }
-
-            intent.setData(uriBuilder.build());
-            activity.startActivityForResult(intent, requestCode);
-            return true;
-        }
     }
 
 }
